[Rust][Windows] Fix frontend runtime lifecycle
Adapt e6b7d7979620caea5f897809dc92239fad96d5eb to v0.29. Restore Windows socket inheritance and adoption, 64-bit listener handles, Ctrl-Break shutdown with forced fallback, and Windows regression coverage.

Preserve the current data-parallel arguments and credential-redacted launch logging. Keep the reviewed Triton repository, pin, and imports unchanged.

+++ rust/src/cmd/src/cli.rs
@@ -458,7 +458,7 @@ impl SharedRuntimeArgs {
     /// inherits an already open HTTP listener from the supervisor process.
     fn into_bootstrapped_config(
         self,
-        listen_fd: i32,
+        listen_fd: u64,
         input_address: String,
         output_address: String,
         coordinator_address: Option<String>,
@@ -648,10 +648,10 @@ fn parse_runtime_args_json(value: &str) -> Result<SharedRuntimeArgs, String> {
 #[derive(Educe, Clone, Args, PartialEq, Eq)]
 #[educe(Debug)]
 pub struct FrontendArgs {
-    /// Inherited listening socket file descriptor passed by the Python
+    /// Inherited listener descriptor or socket handle passed by the Python
     /// supervisor.
     #[arg(long)]
-    pub listen_fd: i32,
+    pub listen_fd: u64,
     /// Frontend input ROUTER socket address that the Python engines will
     /// connect to.
     #[arg(long)]
